Reorganized DictLikeAPI trait
